+++ R/format_vec.R
@@ -3799,9 +3799,12 @@ gt_one_col <- function(x) {
   gt(dplyr::tibble(x = x), auto_align = FALSE, process_md = FALSE)
 }
 
-stop_if_not_vector <- function(x) {
+stop_if_not_vector <- function(x, call = rlang::caller_env()) {
   if (!rlang::is_vector(x)) {
-    cli::cli_abort("The object provided as `x` should be a vector.")
+    cli::cli_abort(
+      "`x` must be a vector, not {.obj_type_friendly {x}}.",
+      call = call
+    )
   }
 }