+++ R/gt-package.R
@@ -23,5 +23,4 @@
 
 
 #' @keywords internal
-#' @aliases gt-package
 "_PACKAGE"