+++ R/helpers.R
@@ -3318,6 +3318,7 @@ cell_text <- function(
   cell_style_structure(name = "cell_text", obj = style_vals)
 }
 
+#' @export
 cell_style_to_html.cell_text <- function(style) {
 
   css <- unclass(style)
@@ -3437,6 +3438,7 @@ cell_fill <- function(
   cell_style_structure("cell_fill", style_vals)
 }
 
+#' @export
 cell_style_to_html.cell_fill <- function(style) {
 
   css <- list()
@@ -3642,6 +3644,7 @@ cell_borders <- function(
   as_style(style = style_list)
 }
 
+#' @export
 cell_style_to_html.cell_border <- function(style) {
 
   css <- unclass(style)