+++ R/tab_create_modify.R
@@ -2418,12 +2418,19 @@ tab_footnote <- function(
   for (loc in locations) {
 
     data <-
-      set_footnote(
-        loc = loc,
-        data = data,
-        footnote = footnote,
-        placement = placement
-      )
+      withCallingHandlers(
+        set_footnote(
+          loc = loc,
+          data = data,
+          footnote = footnote,
+          placement = placement
+      ),
+      error = function(e) {
+        cli::cli_abort(
+          "Can't add footnote {.val {footnote}}.",
+          parent = e
+        )
+      })
   }
 
   data
@@ -2433,6 +2440,7 @@ set_footnote <- function(loc, data, footnote, placement) {
   UseMethod("set_footnote")
 }
 
+#' @export
 set_footnote.cells_title <- function(
     loc,
     data,
@@ -2475,6 +2483,7 @@ set_footnote.cells_title <- function(
   data
 }
 
+#' @export
 set_footnote.cells_stubhead <- function(
     loc,
     data,
@@ -2497,6 +2506,7 @@ set_footnote.cells_stubhead <- function(
   data
 }
 
+#' @export
 set_footnote.cells_column_labels <- function(
     loc,
     data,
@@ -2504,7 +2514,7 @@ set_footnote.cells_column_labels <- function(
     placement
 ) {
 
-  resolved <- resolve_cells_column_labels(data = data, object = loc)
+  resolved <- resolve_cells_column_labels(data = data, object = loc, call = call("cells_column_labels"))
 
   cols <- resolved$columns
 
@@ -2525,6 +2535,7 @@ set_footnote.cells_column_labels <- function(
   data
 }
 
+#' @export
 set_footnote.cells_column_spanners <- function(
     loc,
     data,
@@ -2532,7 +2543,7 @@ set_footnote.cells_column_spanners <- function(
     placement
 ) {
 
-  resolved <- resolve_cells_column_spanners(data = data, object = loc)
+  resolved <- resolve_cells_column_spanners(data = data, object = loc, call = call("cells_column_spanners"))
 
   groups <- resolved$spanners
 
@@ -2551,6 +2562,7 @@ set_footnote.cells_column_spanners <- function(
   data
 }
 
+#' @export
 set_footnote.cells_row_groups <- function(
     loc,
     data,
@@ -2585,6 +2597,7 @@ set_footnote.cells_row_groups <- function(
   data
 }
 
+#' @export
 set_footnote.cells_body <- function(
     loc,
     data,
@@ -2592,7 +2605,7 @@ set_footnote.cells_body <- function(
     placement
 ) {
 
-  resolved <- resolve_cells_body(data = data, object = loc)
+  resolved <- resolve_cells_body(data = data, object = loc, call = call("cells_body"))
 
   rows <- resolved$rows
 
@@ -2613,6 +2626,7 @@ set_footnote.cells_body <- function(
   data
 }
 
+#' @export
 set_footnote.cells_stub <- function(
     loc,
     data,
@@ -2639,6 +2653,7 @@ set_footnote.cells_stub <- function(
   data
 }
 
+#' @export
 set_footnote.cells_summary <- function(
     loc,
     data,
@@ -2655,6 +2670,7 @@ set_footnote.cells_summary <- function(
   )
 }
 
+#' @export
 set_footnote.cells_grand_summary <- function(
     loc,
     data,
@@ -2671,6 +2687,7 @@ set_footnote.cells_grand_summary <- function(
   )
 }
 
+#' @export
 set_footnote.cells_stub_summary <- function(
     loc,
     data,
@@ -2703,6 +2720,7 @@ set_footnote.cells_stub_grand_summary <- function(
   )
 }
 
+#' @export
 set_footnote.cells_source_notes <- function(
     loc,
     data,
@@ -2712,6 +2730,7 @@ set_footnote.cells_source_notes <- function(
   cli::cli_abort("Footnotes cannot be applied to source notes.")
 }
 
+#' @export
 set_footnote.cells_footnotes <- function(
     loc,
     data,
@@ -3502,6 +3521,7 @@ set_style <- function(loc, data, style) {
   UseMethod("set_style")
 }
 
+#' @export
 set_style.cells_title <- function(loc, data, style) {
 
   title_components <- rlang::eval_tidy(loc$groups)
@@ -3537,6 +3557,7 @@ set_style.cells_title <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_stubhead <- function(loc, data, style) {
 
   data <-
@@ -3553,6 +3574,7 @@ set_style.cells_stubhead <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_column_labels <- function(loc, data, style) {
 
   resolved <- resolve_cells_column_labels(data = data, object = loc)
@@ -3575,6 +3597,7 @@ set_style.cells_column_labels <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_column_spanners <- function(loc, data, style) {
 
   resolved <- resolve_cells_column_spanners(data = data, object = loc)
@@ -3595,6 +3618,7 @@ set_style.cells_column_spanners <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_row_groups <- function(loc, data, style) {
 
   row_groups <- dt_row_groups_get(data = data)
@@ -3623,6 +3647,7 @@ set_style.cells_row_groups <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_body <- function(loc, data, style) {
 
   resolved <- resolve_cells_body(data = data, object = loc)
@@ -3645,6 +3670,7 @@ set_style.cells_body <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_stub <- function(loc, data, style) {
 
   resolved <- resolve_cells_stub(data = data, object = loc)
@@ -3665,6 +3691,7 @@ set_style.cells_stub <- function(loc, data, style) {
   data
 }
 
+#' @export
 set_style.cells_summary <- function(loc, data, style) {
 
   add_summary_location_row(
@@ -3675,6 +3702,7 @@ set_style.cells_summary <- function(loc, data, style) {
   )
 }
 
+#' @export
 set_style.cells_grand_summary <- function(loc, data, style) {
 
   add_grand_summary_location_row(
@@ -3685,6 +3713,7 @@ set_style.cells_grand_summary <- function(loc, data, style) {
   )
 }
 
+#' @export
 set_style.cells_stub_summary <- function(loc, data, style) {
 
   add_summary_location_row(
@@ -3695,6 +3724,7 @@ set_style.cells_stub_summary <- function(loc, data, style) {
   )
 }
 
+#' @export
 set_style.cells_stub_grand_summary <- function(loc, data, style) {
 
   add_grand_summary_location_row(
@@ -3705,6 +3735,7 @@ set_style.cells_stub_grand_summary <- function(loc, data, style) {
   )
 }
 
+#' @export
 set_style.cells_footnotes <- function(loc, data, style) {
 
   data <-
@@ -3719,6 +3750,7 @@ set_style.cells_footnotes <- function(loc, data, style) {
     )
 }
 
+#' @export
 set_style.cells_source_notes <- function(loc, data, style) {
 
   data <-