+++ tests/testthat/test-fmt_currency.R
@@ -865,14 +865,18 @@ test_that("The `fmt_currency()` fn can render in the Indian numbering system", {
       "&#8377;0.00", "NA", "&#8377;Inf Cr", "(&#8377;Inf) Cr"
     )
   )
-  expect_warning(
-    expect_equal(
-      (tab %>%
-         fmt_currency(columns = num, suffixing = TRUE, system = "ind") %>%
-         render_formats_test(context = "html"))[["num"]],
-      (tab %>%
-         fmt_currency(columns = num, suffixing = TRUE, scale_by = 200, system = "ind") %>%
-         render_formats_test(context = "html"))[["num"]]
-    )
+
+  expect_no_warning(
+    compared_tab <- tab %>%
+      fmt_currency(columns = num, suffixing = TRUE, system = "ind")
+  )
+  # scale_by warning
+  expect_snapshot(
+    expected_tab <- tab %>%
+      fmt_currency(columns = num, suffixing = TRUE, scale_by = 200, system = "ind")
+  )
+  expect_equal(
+    render_formats_test(compared_tab, context = "html")[["num"]],
+    render_formats_test(expected_tab, context = "html")[["num"]]
   )
 })